Refactor main.cpp to guarantee the binary consistency of image.
As the readdir doesn't guarantee the order, the creating order may be
different in various enviroments. Here using BFS instead of DFS, and
sort the sequence, the creating order can be guaranteed. This is good
for binary consistency.

Also here the timestamp is not set defaultly, because it will make a
little bit different between images.