Automatically grab FS size from image size
Fixes #20
Fixes #29
Fixes #14

The CLI required the user to pass in the FS size, but it seems most people
neglected to do this and the default 0x10000 was used, causing LittleFS to
truncate most images.

Now require the -s size option on creation, and auto-read the size when doing
a list or update operation.  Users can still manually code in a size from the
command like if needed.